vlm_infer/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
tqdm>=4.65.0
rich>=13.0.0

# API clients
openai>=1.0.0
requests>=2.31.0

# Data processing
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: for better notebook support
ipython>=8.0.0
jupyter>=1.0.0

# Development tools
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.1.0  # Code linting
mypy>=1.5.0   # Type checking 

# For .env file handling
python-dotenv>=1.0.0 